Split order page test and fix
